updates Pathfinder mode types to add an "IDE" mode
